idp_oidc: handle interaction betwen A2_EMAIL_IS_UNIQUE and STRATEGY_CREATE (fixes #19974)
It should limit the possibity of duplicate accounts BUT as the email field is
not really unique at the database level, we cannot really garantee it.

To improve on the current state we should dynamically create unique constraints
on the mail field based on OU.email_is_unique or A2_EMAIL_IS_UNIQUE parameters.
That's not in the current scope.