manager: remove All ou choice from ou selector on user's roles view (fixes #20519)
User's roles view is special, the choice all ou modify the UI so that it shows
all roles of the user and only them, whatever their OU but it do not show
all roles from all OUs like on the role listing view. To prevent
misunderstanding from the user on the effect of this choice, we change it
for the empty string as before.