manager: send new email in the email change verification mail (fixes #20564)
Use of a ModelForm keeping the original email field for the
UserChangeEmailForm makes keeping the original email value after clean()
is called impossible, as clean() is also responsible of transfering
value from the form into the model instance.

We keep using a ModelForm but we use a new field not present in the
model to get the new email and we override the save() method so that the
behaviour of sending the validation mail is kept inside the form and not
in the view. Only the call to the manager's hook
manager-change-email-request is kept in the view.