tests: disable cache decorators by default (fixes #22227)
It makes behavior of tests erratic as it incurs side effects between
tests and even inside of tests. For example, caching of the OU count for
short periods is not harmful for day to day use as creating new OUs are
rare events but during tests it makes reasoning about code behavior
really difficult.